practica1/backend2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
boto3==1.29.7
Pillow==10.1.0
aiofiles==23.2.1
